requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment management
python-dotenv==1.0.0

# Data processing and export
pandas==2.1.4
openpyxl==3.1.2
fpdf2==2.7.6

# HTTP client for external APIs
httpx==0.25.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Optional: For advanced features
celery==5.3.4
redis==5.0.1
email-validator==2.1.0
jinja2==3.1.2
python-slugify==8.0.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.39.2
